- List viewmodel - Repository - Navigation - DI implemented - Dependencies added

+++ app/src/main/java/com/android/objectmanagerapp/data/source/local/entity/RelationEntity.kt
@@ -16,7 +16,7 @@ import androidx.room.ForeignKey
     ]
 )
 data class RelationEntity(
-    val parentObjectId: Long,
-    val childObjectId: Long,
+    val parentObjectId: Int,
+    val childObjectId: Int,
     val relationshipType: String
 )
